test(cli): SMI-4474 split recommend.test.ts + add loadStoredAccessToken stub
The vi.mock of @skillsmith/core in packages/cli/tests/recommend.test.ts
didn't stub loadStoredAccessToken — after the SMI-4474 wiring, the
recommend command imports it, calling unmocked threw and crashed the
action handler before console.log fired (31 CI failures).

Adding the stub required modifying recommend.test.ts, which already
exceeded the 500-line standard at 1012 lines. The pre-commit hook
correctly refuses any modification to a file over 500 lines, so the
file is split into 3 sibling files and a pure-fixtures helper:

  - recommend.test.ts (~431) — registration / analyzer / API / output
  - recommend.errors.test.ts (~304) — errors / limit / trust tier
  - recommend.filters.test.ts (~301) — stack / role / exports
  - recommend.test-helpers.ts (~98) — shared fixtures (no mocks; vi.hoisted
    can't be exported, so each test file owns its own mocks block)

All 545 CLI tests pass. The loadStoredAccessToken stub returns null in
each file's mock since none of these tests exercise the JWT path —
behavior is identical to anonymous CLI calls.

Linear: SMI-4474
